+++ email/email_globals.h
@@ -39,7 +39,7 @@ extern struct ListHead Ignore;              ///< List of header patterns to igno
 extern struct RegexList NoSpamList;         ///< List of regexes to whitelist non-spam emails
 extern struct ReplaceList SpamList;         ///< List of regexes and patterns to match spam emails
 extern struct ListHead UnIgnore;            ///< List of header patterns to unignore (see)
-extern struct ListHead MailToAllow;         ///< List of permitted fields in a mailto: uri
+extern struct ListHead MailToAllow;         ///< List of permitted fields in a mailto: url
 extern struct Hash *AutoSubscribeCache;     ///< Hash table of auto-subscribed mailing lists
 extern struct RegexList UnSubscribedLists;  ///< List of regexes to blacklist false matches in SubscribedLists
 extern struct RegexList MailLists;          ///< List of regexes to match mailing lists